Main menu selection. Using whiptail to make the selection. Basic functionality implemented. TODO: add the rest.
At this point, we are startiung to display the options using a pretty
menu instead of the plain bash prompt. Well, for power-users, Bash will
suffice and even we preffer a bash prompt. But, we need to be aware that
the purpose of all this is to make the live a bit easier for the users.